test client: harden the remaining rough edges
- Handlers are copied out under the shared mutex and invoked unlocked;
  a handler calling back into the client (send*, set*MessageHandler)
  would otherwise deadlock on the writer-queued shared mutex.
- The binary service-call and fetch-asset parsers validate frame and
  embedded-string lengths before parsing — and before consuming the
  fulfilled slot, so a malformed frame is dropped rather than causing
  an out-of-bounds read or burning the waiter.
- sendText/sendBinary throw a clear error when disconnected instead of
  websocketpp's opaque get_con_from_hdl exception.
- connect()'s future is settled with an exception on connection
  failure (new fail-handler hook), so a refused connection fails fast
  and diagnosably instead of timing out.
- ChannelId is u32, matching the wire type (and what from_json already
  read); the SDK's internal u64 channel IDs never cross the protocol.

+++ tests/client/protocol_types.hpp
@@ -11,8 +11,9 @@
 
 namespace foxglove::test {
 
-// Protocol type aliases
-using ChannelId = uint64_t;
+// Protocol type aliases. Channel IDs are u32 on the wire (and read as such
+// in from_json), even though the SDK uses u64 channel IDs internally.
+using ChannelId = uint32_t;
 using SubscriptionId = uint32_t;
 using ClientChannelId = uint32_t;
 using ServiceId = uint32_t;
